# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 晟風工程 ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: 列出所有報價單
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 報價單列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: 新增報價單
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteInput"
      responses:
        "201":
          description: 新增成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: 取得報價單詳細資料
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 報價單資料
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: 找不到報價單
    patch:
      operationId: updateQuote
      tags: [quotes]
      summary: 更新報價單
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteUpdate"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: 找不到報價單
    delete:
      operationId: deleteQuote
      tags: [quotes]
      summary: 刪除報價單
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 刪除成功
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: 列出所有收款紀錄
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: 收款紀錄列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: 新增收款紀錄
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: 新增成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: 取得收款紀錄
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 收款紀錄
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: 找不到收款紀錄
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: 更新收款紀錄
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: 刪除收款紀錄
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 刪除成功
  /warranties:
    get:
      operationId: listWarranties
      tags: [warranties]
      summary: 列出所有保固資料
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: 保固列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warranty"
    post:
      operationId: createWarranty
      tags: [warranties]
      summary: 新增保固資料
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarrantyInput"
      responses:
        "201":
          description: 新增成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warranty"
  /warranties/{id}:
    get:
      operationId: getWarranty
      tags: [warranties]
      summary: 取得保固資料
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 保固資料
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warranty"
        "404":
          description: 找不到保固資料
    patch:
      operationId: updateWarranty
      tags: [warranties]
      summary: 更新保固資料
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarrantyUpdate"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warranty"
    delete:
      operationId: deleteWarranty
      tags: [warranties]
      summary: 刪除保固資料
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 刪除成功
...
components:
  schemas:
...
    Quote:
      type: object
      required: [id, customerId, title, amount, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        customerName: { type: ["string", "null"] }
        title: { type: string }
        description: { type: ["string", "null"] }
        amount: { type: number }
        discountAmount: { type: ["number", "null"] }
        finalAmount: { type: ["number", "null"] }
        status: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    QuoteInput:
      type: object
      required: [customerId, title, amount, status]
      properties:
        customerId: { type: integer }
        title: { type: string, minLength: 1 }
        description: { type: string }
        amount: { type: number }
        discountAmount: { type: number }
        finalAmount: { type: number }
        status: { type: string }
        notes: { type: string }
    QuoteUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        amount: { type: number }
        discountAmount: { type: number }
        finalAmount: { type: number }
        status: { type: string }
        notes: { type: string }
...
    Payment:
      type: object
      required: [id, customerId, amount, paymentDate, createdAt]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        customerName: { type: ["string", "null"] }
        quoteId: { type: ["integer", "null"] }
        workOrderId: { type: ["integer", "null"] }
        amount: { type: number }
        paymentDate: { type: string }
        paymentMethod: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    PaymentInput:
      type: object
      required: [customerId, amount, paymentDate]
      properties:
        customerId: { type: integer }
        quoteId: { type: integer }
        workOrderId: { type: integer }
        amount: { type: number }
        paymentDate: { type: string }
        paymentMethod: { type: string }
        notes: { type: string }
    PaymentUpdate:
      type: object
      properties:
        amount: { type: number }
        paymentDate: { type: string }
        paymentMethod: { type: string }
        notes: { type: string }
    Warranty:
      type: object
      required: [id, customerId, startDate, endDate, createdAt]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        customerName: { type: ["string", "null"] }
        acUnitId: { type: ["integer", "null"] }
        startDate: { type: string }
        endDate: { type: string }
        description: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    WarrantyInput:
      type: object
      required: [customerId, startDate, endDate]
      properties:
        customerId: { type: integer }
        acUnitId: { type: integer }
        startDate: { type: string }
        endDate: { type: string }
        description: { type: string }
        notes: { type: string }
    WarrantyUpdate:
      type: object
      properties:
        startDate: { type: string }
        endDate: { type: string }
        description: { type: string }
        notes: { type: string }
...