# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 晟風工程 ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface Payment {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  quoteId?: number | null;
  /** @nullable */
  workOrderId?: number | null;
  amount: number;
  paymentDate: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}