# lib/api-zod/src/generated/types/quoteInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 晟風工程 ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteInput {
  customerId: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  amount: number;
  discountAmount?: number;
  finalAmount?: number;
  status: string;
  notes?: string;
}